updated from class
