+++ crud_dinosaurs/cryptids.json
@@ -1,1 +1,1 @@
-[{"name":"Loch Ness Monster","img_url":"https://www.history.com/.image/t_share/MTU3ODc5MDg3MjQzNjAxMjI1/loch-ness-monster.jpg"},{"name":"Big Foot","img_url":"https://res.cloudinary.com/simpleview/image/upload/crm/eurekaca/patterson.gimlin0-b5b6c02e5056a36_b5b6c19f-5056-a36a-096e6e95925c7427.jpg"},{"name":"Batsquatch","img_url":"https://vignette.wikia.nocookie.net/cryptidz/images/c/c7/Batsquatch.jpg/revision/latest?cb=20100912164353"},{"name":"Sharlie (Slimy Slim)","img_url":"https://vignette.wikia.nocookie.net/cryptidz/images/6/6c/Swimsharlie.jpg"},{"name":"The Kraken","img_url":"https://upload.wikimedia.org/wikipedia/commons/thumb/9/9d/Colossal_octopus_by_Pierre_Denys_de_Montfort.jpg/220px-Colossal_octopus_by_Pierre_Denys_de_Montfort.jpg"}]
+[{"name":"nessie","img_url":"https://www.history.com/.image/t_share/MTU3ODc5MDg3MjQzNjAxMjI1/loch-ness-monster.jpg"},{"name":"Big Foot","img_url":"https://res.cloudinary.com/simpleview/image/upload/crm/eurekaca/patterson.gimlin0-b5b6c02e5056a36_b5b6c19f-5056-a36a-096e6e95925c7427.jpg"},{"name":"batsquatch","img_url":"https://vignette.wikia.nocookie.net/cryptidz/images/c/c7/Batsquatch.jpg/revision/latest?cb=20100912164353"},{"name":"Sharlie (Slimy Slim)","img_url":"https://vignette.wikia.nocookie.net/cryptidz/images/6/6c/Swimsharlie.jpg"},{"name":"The Kraken","img_url":"https://upload.wikimedia.org/wikipedia/commons/thumb/9/9d/Colossal_octopus_by_Pierre_Denys_de_Montfort.jpg/220px-Colossal_octopus_by_Pierre_Denys_de_Montfort.jpg"}]
