+++ crud_dinosaurs/dinosaurs.json
@@ -1,1 +1,1 @@
-[{"name":"Littlefoot","type":"Apatosaurus"},{"name":"Cera","type":"Triceratops"},{"name":"Ducky","type":"Saurolophus"},{"name":"Petrie","type":"Pteranodon"},{"name":"Spike","type":"Stegosaurus"},{"type":"test","name":"testy"},{"type":"t-rex","name":"rexy"}]
+[{"name":"Littlefoot","type":"Apatosaurus"},{"name":"Cera","type":"Triceratops"},{"name":"Ducky","type":"Saurolophus"},{"name":"Petrie","type":"Pteranodon"},{"name":"Spike","type":"Stegosaurus"},{"type":"test","name":"testy"},{"type":"t-rex","name":"rexy"},{"type":"classTest","name":"classTest"},{"type":"classTest","name":"classTest"},{"type":"newtest","name":"newtest"}]
