+++ crud_dinosaurs/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "ejs": "^3.0.1",
     "express": "^4.17.1",
-    "express-ejs-layouts": "^2.5.0"
+    "express-ejs-layouts": "^2.5.0",
+    "method-override": "^3.0.0"
   }
 }
