+++ crud_dinosaurs/views/layout.ejs
@@ -11,11 +11,8 @@
         <ul>
             <li><a href='/dinosaurs'>View All Dinosaurs</a></li>
             <li><a href='/dinosaurs/new'>Add New Dinosaur</a></li>
-            <li><a href='/dinosaurs/0'>Show 1st Dinosaur</a></li>
             <li><a href='/cryptids'>View All Cryptids</a></li>
-            <li><a href='/cryptids/0'>Show 1st Cryptid</a></li>
             <li><a href='/cryptids/new'>Add New Cryptid</a></li>
-            <li><a href='/cryptids/edit/0'>Edit 1st Cryptid</a></li>
         </ul>
     </nav>
     <%- body %>
